src/main/dataCollector/sourceTwo.py
import os
import argparse
from __future__ import print_function
import torch
import torchrec
import torch.distributed as dist
import torch.nn as nn
import torch.nn.functional as F
import torch.optim as optim
from torchvision import datasets, transforms
from torch.optim.lr_scheduler import StepL
import numpy as np

class SourceTwo(nn.Module):
    
    """
    
    SourceTwo.py: Public clothing datasets to recognize individual articles of clothing.

    -------------------------------------------------------------------------------------------

    Functions:

    """
    def __init__(self, ):
        pass
